test_api/requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.3
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
openpyxl==2.6.3
Pillow==6.1.0
pluggy==0.13.1
py==1.8.0
PyMySQL==0.9.3
pytesseract==0.2.7
pytest==4.5.0
pytest-html==1.19.0
pytest-metadata==1.8.0
PyTestReport==0.1.9.3
PyYAML==5.3
requests==2.22.0
selenium==3.141.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
